[add]http通信のコンポーネントを作った。
