+++ lib/view/components/constant.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:http/http.dart' as http;  // http
+import 'dart:convert';  // json
+
 
 class Constant {
   static const Color main = Color(0xFF82E2E4);
@@ -148,3 +151,75 @@ class _HorizontalScrollingBackgroundState extends State<HorizontalScrollingBackg
   }
 }
 */
+
+// Httpリクエストを投げるクラス
+// 
+class HttpToServer {  // HttpLib
+  // インスタンス変数
+  // URLとかポートとかプロトコルとか
+  static const String serverIP = "127.0.0.1";  // "127.0.0.1""10.200.0.82""tidalhip.local""10.200.0.115"10.25.10.10710.200.0.163
+  static const String server_port = "5000";
+  static const String protocol = "http";
+  static baseUrl() {  // 鯖のURLを設定
+    return protocol + "://" + serverIP + ":" + server_port;
+  }
+
+  // コンストラクタってなに？？
+  HttpToServer();
+
+  // "/send_userInfo", "POST", items.userInfo
+  // HTTP リクエストを送信する関数 
+  static Future<List> httpReq(String path_para, String method, Map<String, dynamic> body) async {
+    //header
+    var headersList = {
+      'Accept': '*/*',
+      'User-Agent': 'Thunder Client (https://www.thunderclient.com)',
+      'Content-Type': 'application/json',
+    };
+    // "Access-Control-Allow-Origin": "*"
+
+    // リクエスト作成
+    var req = http.Request("POST", Uri.parse(baseUrl() + path_para));  // HTTPリクエストメソッドの種類とuriから
+      // debugPrint(req.toString());
+    req.headers.addAll(headersList);  // header情報を追加
+      // debugPrint(req.toString());
+    req.body = json.encode(body);  // bodyをjson形式に変換
+      // debugPrint(req.toString());
+
+    try {
+
+      // HTTPリクエストを送信。 seconds: 5 で指定した秒数応答がなかったらタイムアウトで例外を発生させる
+      var res = await req.send();//.timeout(const Duration(seconds: 5));
+      debugPrint(res.toString());
+      debugPrint("qawsedrftyhuji");
+      // レスポンスをストリームから文字列に変換して保存
+      final resBody = await res.stream.bytesToString();
+
+      // ステータスコードが正常ならtrueと内容を返す
+      if (res.statusCode >= 200 && res.statusCode < 300) {
+        debugPrint([true, resBody].toString());
+        return [true, resBody];
+      } else {
+        debugPrint([false, res.statusCode.toString(), resBody].toString());
+        return [false, res.statusCode.toString(), resBody];
+      }
+
+    } catch (e) {  // タイムアウトしたとき。
+
+      debugPrint("exception error: " + e.toString());
+      debugPrint([false, "おうとうないよ；；"].toString());
+      return [false, "おうとうないよ；；"];
+
+    }
+
+
+    // if (response.statusCode == 200) {
+    //   // レスポンスが成功した場合
+    //   final data = json.decode(response.body);
+      
+    // } else {
+    //   // レスポンスがエラーの場合
+     
+    // }
+  }
+}
